+++ lib/grape/middleware/versioner/header.rb
@@ -135,7 +135,7 @@ module Grape
         end
 
         def version_not_found!(media_types)
-          return unless media_types.all? { |media_type| media_type&.version && versions && !versions.include?(media_type.version) }
+          return unless media_types.all? { |media_type| media_type&.version && versions&.none? { |version| version.casecmp?(media_type.version) } }
 
           invalid_version_header!('API version not found.')
         end